requirements.txt
psycopg2-binary==2.9.9
SQLAlchemy==2.0.27
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.2.0
fastapi==0.135.3
uvicorn==0.43.0
python-telegram-bot==20.7
python-telegram-bot==20.7
requests==2.31.0
